[fix] Revert "[fix] change the field type of disk size from int to long [resource gather(in the BalanceTaskSummary)]"
This reverts commit 234ad01e

+++ FS_Proto/src/main/java/com/bonree/brfs/common/rebalance/route/NormalRouteInterface.java
@@ -56,7 +56,7 @@ public interface NormalRouteInterface {
      *
      * @return
      */
-    Map<String, Long> getRoutes();
+    Map<String, Integer> getRoutes();
 
     /**
      * 获取二级serverid与一级serverid的关系
